Added new "INTERNAL" error type to report compiler errors. Decided against puttint errors in hashing function as it is suppose to be interchangable. So hash table expects a 0 and then reports the error. Started working on replacing all error messages to use the new diagnostic system. Cleaned up main with cleanup function that checks for null values before freeing. Removed compiler stages in context.h that are not needed yet. Fixed mismatch with return value of fnv1a function . Changed init keyword table function to non-static. Added compiler context into hash table code.

+++ fnv-1a.c
@@ -1,5 +1,6 @@
 #include "diagnostic.h"
 #include "context.h"
+#include "fnv-1a.h"
 
 #include <stddef.h>
 #include <stdint.h>
@@ -15,14 +16,13 @@ TODO(malac0da):
 #define FNV_PRIME 16777619
 #define FNV_OFFSET_BASIS 2166136261
 
-uint32_t fnv1a(const char *key)
+unsigned long fnv1a(const char *key)
 {
     if (!key) {                 // TODO(malac0da): Report error for null key
-        //report_error("FNV-1a: NULL key provided.\n");
         return 0;
     }
     // FNV-1a offset basis
-    uint32_t hash = FNV_OFFSET_BASIS;
+    unsigned long hash = FNV_OFFSET_BASIS;
     // strlen ensures that the string is null-terminated. So checking for it is redundant
     size_t length = strlen(key);
